requirements.txt


numpy==1.23.5
Pillow==9.4.0
fastapi==0.103.1
uvicorn==0.23.2
python-multipart==0.0.6

